· 스냅샷 핀 충돌: BE-4b 의 `roster_auto_confirmed_at` 추가와 이 브랜치의 이의 모델 제거가
  **한 파일에 함께** 들어왔다. 어느 한쪽 해시를 그대로 쓰면 병합 결과와 달라 CI 가 깨지므로
  **이력 주석은 둘 다 남기고 해시는 병합본으로 다시 계산**했다.
· `count(*)` 는 bigint 다. integer 로 받으면 넘칠 뿐 아니라, 타입이 실제 반환형과 다르면
  읽는 사람이 무엇이 오는지 잘못 안다.

검증: 이의 2행을 심은 DB 에서 1회차 `행 수 = 2` NOTICE 후 제거, 2회차 "이미 없다" no-op.
스냅샷 가드 포함 유닛 13/13, expand-contract 게이트 4건 전부 수락(rc=0).

+++ apps/v1_api/prisma/migrations/20260903170000_v1_drop_league_match_disputes/migration.sql
@@ -17,7 +17,9 @@
 -- 이미 배포돼 있다는 것이 그 전제이고, 위 grep 0 이 그 근거다.
 DO $$
 DECLARE
-  dispute_count integer;
+  -- `count(*)` 는 bigint 다. integer 로 받으면 20억 행에서 넘치는데, 그보다 중요한 건
+  -- 타입이 실제 반환형과 다르면 읽는 사람이 무엇이 오는지 잘못 안다는 것이다.
+  dispute_count bigint;
 BEGIN
   IF to_regclass('public.v1_league_match_disputes') IS NULL THEN
     RAISE NOTICE 'task166-dispute-drop: 테이블이 이미 없다 — 건너뛴다';
